取り崩し処理：選択ボタン非活性時スタイル適用

+++ ToDoApp_01/src/main/java/com/example/demo/form/SavingPurposeForm.java
@@ -1,5 +1,7 @@
 package com.example.demo.form;
 
+import java.math.BigDecimal;
+
 import jakarta.validation.constraints.DecimalMin;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
@@ -9,5 +11,5 @@ public class SavingPurposeForm {
 	private String name;
 	@NotNull(message = "amount is required")
     @DecimalMin(value = "0.0", inclusive = false, message = "amount must be greater than zero")
-	private Double neededAmount;
+	private BigDecimal neededAmount;
 }
